Bug fixes +Endpoints secured

+++ Sonstiges/htdocs/restApi/rest/deleteChallenges.php
@@ -10,6 +10,7 @@ header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers
 include_once '../config/database.php';
 include_once '../model/Challenge.php';
 include_once '../queries/Query.php';
+include_once '../../vendor/autoload.php';
 
 // get database connection
 $database = new Database();
@@ -17,12 +18,15 @@ $db = $database->getConnection();
 $query = new Query($db);
 $data = file_get_contents('php://input');
 $data = json_decode($data, true);
-$year = $data['object'];
+$year = $data['year'];
+$idToken = $data['idtoken'];
 
-if($query->deleteChallengesInYear($year)){
-    header('HTTP/1.1 200 OK');
-}
-
-else{
-    header('HTTP/1.1 500 Internal Server Error');
+if($query->getUserRights($idToken) != 'schramm'){
+    http_response_code(401);
+}else {
+    if ($query->deleteChallengesInYear($year)) {
+        header('HTTP/1.1 200 OK');
+    } else {
+        header('HTTP/1.1 500 Internal Server Error');
+    }
 }
